requirements.txt
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp~=3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
attrs==24.3.0
boto3~=1.35.97
botocore~=1.35.99
certifi==2024.12.14
charset-normalizer==3.4.1
ffmpeg==1.4
frozenlist==1.5.0
grpcio==1.69.0
grpcio-tools==1.69.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
propcache==0.2.1
protobuf==5.29.3
psycopg2==2.9.10
pydub==0.25.1
pymediainfo==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.1
setuptools==75.8.0
six==1.17.0
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
openai~=1.62.0
anthropic~=0.45.2